# server/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.1 (beta)
  title: api-rb
  description: RB
host: 'api.rb.local:1234'
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /pdisk-reservation/calendar/{weekDay}/{turnNum}:
    get:
      tags:
        - pdiskreservation
      parameters:
        - $ref: '#/parameters/authToken'
        - $ref: '#/parameters/pdiskreservation.turnWeekDay.path'
        - $ref: '#/parameters/pdiskreservation.turnNum.path'
      responses:
        '200':
          schema:
            type: integer
          description: Count
        '401':
          description: Unauthorized
        '403':
          description: Forbbiden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
parameters:
...
  authToken:
    name: AuthToken
    description: for authentication
    required: false
    in: header
    type: string
...
  pdiskreservation.turnWeekDay.path:
    name: weekDay
    description: turnWeekDay
    required: true
    in: path
    type: string
...
  pdiskreservation.turnNum.path:
    name: turnNum
    description: turnNum
    required: true
    in: path
    type: integer
...